refactor: add TagService and TodoService, WIP
